LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/navigating-across-documents/about-srcdoc-navigation-blocked-expected.txt



FAIL Navigations to about:srcdoc via window.location are blocked. assert_equals: expected "blocked" but got "allowed"
FAIL Navigations to about:srcdoc via window.open() are blocked. assert_equals: expected "blocked" but got "allowed"
